Project modifications - fixed size_t to int cast warnings - fixed include paths for libusb liblo - removed some lua binding methods that cause build errors (need to check them again to see if they should be added back) - removed LibraryEditor (agaiN!)

+++ Source/Core/CtrlrFontManager.cpp
@@ -284,9 +284,9 @@ const String CtrlrFontManager::getStringFromFont (const Font &_font)
 	StringArray fontProps;
 	fontProps.add (font.getTypefaceName());
 	fontProps.add (String(font.getHeight()));
-	fontProps.add (String(font.isBold()));
-	fontProps.add (String(font.isItalic()));
-	fontProps.add (String(font.isUnderlined()));
+	fontProps.add (String((int)font.isBold()));
+	fontProps.add (String((int)font.isItalic()));
+	fontProps.add (String((int)font.isUnderlined()));
 	fontProps.add (String(font.getExtraKerningFactor()));
 	fontProps.add (String(font.getHorizontalScale()));
 	fontProps.add (String((uint8)getFontSetEnum (font)));
